Add diagnostic notation
