+++ src/library.h
@@ -6,7 +6,8 @@ typedef unsigned long long uint64_t;
 #define DMAX_DEPTH 20
 #define DFAIL -128
 
-#define FOREACH_PHASE(FN) FN(BEGIN), FN(ITEM), FN(FINISH), FN(ERROR)
+#define FOREACH_PHASE(FN) \
+  FN(BEGIN), FN(BEFORE_ITEM), FN(AFTER_ITEM), FN(FINISH), FN(ERROR)
 
 // Two interoperable enums; States overlaps MT except for FAIL.
 #define FOREACH_MT(FN) \
